media-analysis-app/requirements.txt
streamlit
requests
lxml
deep-translator
pydub
python-dotenv
concurrent.futures
langdetect
groq
math
ffmpeg